requirements.txt
Django>=5.2.0
djangorestframework>=3.14.0
django-cors-headers>=4.3.0
python-dotenv>=1.0.0
Pillow>=10.0.0
requests>=2.31.0 